backend/requirements.txt
Django>=4.2
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
Pillow>=9.1.0
django-cors-headers>=3.13.0
drf-yasg>=1.21.5
djangorestframework-simplejwt>=5.2.2
django-environ>=0.9.0
python-decouple>=3.7
django-filter>=22.1


# Testing y desarrollo
pytest>=7.1.2
pytest-django>=4.5.2
black>=23.1a0
flake8>=6.0.0
